+++ docs/changes.rst
@@ -13,6 +13,10 @@ Pending
 * Stayed on top of pre-commit hook updates.
 * Added :doc:`architecture documentation <architecture>` to help
   on-board new contributors.
+* Removed the static file path validation check in
+  :class:`StaticFilesPanel <debug_toolbar.panels.staticfiles.StaticFilesPanel>`
+  since that check is made redundant by a similar check in Django 4.0 and
+  later.
 
 4.3.0 (2024-02-01)
 ------------------